Add Website Link to #modal-footer
Added a website link to the #modal-footer for each item. The most noteworthy aspect of this addition is the addition of an .on('click') event to the button element with a corresponding .off('click') event to ensure that the correct link information is passed to the button based on the item that was clicked on.